Integret_with_IDE_MCP/requirements.txt
fastmcp>=2.10.6,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
langchain-mcp-adapters>=0.1.9,<1.0.0
langgraph-prebuilt>=0.5.2,<1.0.0
langchain-groq>=0.3.6,<1.0.0
langgraph>=0.5.4,<0.6.0
langchain-google-genai